+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-All notable changes to `laravel-countries` will be documented in this file
+All notable changes to `laravel-country` will be documented in this file
 
-## 1.0.0 - 2021-08-7
+## 1.0.1 - 2021-08-08
+### Fixed
+- Laravel CountryServiceProvider
+
+## 1.0.0 - 2021-08-07
 ### Added
 - Country migration, seeders and model